requirements.txt
arrow==1.2.3
asgiref==3.5.2
beautifulsoup4==4.11.1
blessed==1.20.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
contourpy==1.0.7
cycler==0.11.0
Django==4.1.5
django-crispy-forms==1.14.0
django-paginate==0.3.1
django-pagination==1.0.7
django-picklefield==3.1
django-q==1.3.9
fonttools==4.39.3
gunicorn==20.1.0
idna==3.3
image==1.5.33
install==1.3.5
kiwisolver==1.4.4
lxml==4.8.0
matplotlib==3.7.1
mod-wsgi==4.9.4
numpy==1.24.2
opencv-contrib-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
pyparsing==3.0.9
pysqlite3==0.5.0
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
requests==2.27.1
schedule==1.1.0
six==1.16.0
smartcrop==0.3.3
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.6
